Animations are displaying correctly with async module
Basic desk operations do not use MultiLock as much as they
used to. The only place where the MultiLock is used now is
when an event is triggered from a model. MultiLock is locked
before an event is triggered and unlocked when all handlers
are finished. If there is no handler than the MultiLock is
unlocked immediately.

Triggering an event is currently the only place where the
MultiLock is necessary because we do not know how many
handlers the event has and we need to wait for all of them
to finis. In cases where we know how many calls will be made
the async library is used. If calls should be called in series
(eg. when activating board phases) then async.series() is used.
If calls should be made in parallel (eg. activating all
conveyors) then async.parallel() is used.

This should significantly lower the number of errors introduced
by bad locking and unlocking.

What should work:
- Entity animations (conveyors, turners, ...)
- Robot animations (moving, rotating)
What does not work yet:
- Faling into a hole

Also one testing board with turners was added.