Basic automated tests with Mocha framework
Tests are stored in the `test/` directory and have file extension
`.test.coffee`. Tests are written in coffee-script and compiled into
javascript using `make tests`. After that display file test/index.html
in a browser to run tests.

Tests can be currently run only in browser not in node.
That is because I was not able to get requirejs to work
with node. Unfortunately our project relies on requirejs heavily.