Introduced concept of effects
Tile does not affect other entities directly but creates an effect
instead. These effects can be chained together.
When all tiles are activated, all effects are checked and
some of them may be removed (eg. robot moving against a wall)
and those effects that are remaining are applied.

This currently works only for conveyors and only initial effects
of chains are applied. In future more sophisticated system will
be implemented.