Spine models can be used server side
This required some hacks in browser. On server, it is necessary
to load async, Spine and underscore with require. But it is not
possible in browser. For that reason there are three *-shim.js
files (one for each of these libraries).

Take for example the async-shim.js file. It is mapped to 'async'
in requirejs config and it just exports the async variable. The
async variable is global in the browser so it is not necessary.
until we load the module server side. On the server there
are no global variables and async module must be loaded in each
file using require('async'). The shim files are not used on the server
so the actual async library from node is loaded instead. This way
the exactly same module files can be used both client side and
server side.